Get Started view on login + Firebase
User data is stored in firebase when account is created.

Asks new users to choose a username and profile picture (optional) the first time they login.

It works by checking if the user has a username as I dont want them to proceed without one. This is the reason im not using .additionalUserInfo.isNewUser because it can be bypassed by restarting app